v3

+++ aggregator/src/main/java/com/twilight/aggregator/model/Token.java
@@ -1,18 +1,17 @@
 package com.twilight.aggregator.model;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
 
 import lombok.Data;
 
 @Data
 public class Token implements Serializable {
     private static final long serialVersionUID = 1L;
-    
+
     private String tokenAddress;
     private Long tokenId;
-    private BigDecimal tokenPriceUsd;
+    private double tokenPriceUsd;
     private boolean buyOrSell;
-    private BigDecimal amount;
+    private double amount;
     private String fromAddress;
-} 
+}
